requirements.txt
appdirs==1.4.4
argon2-cffi==20.1.0
async-generator==1.10
atomicwrites==1.1.5
attrs==19.3.0
backcall==0.1.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
bleach==3.1.1
blinker==1.4
blis==0.4.1
catalogue==1.0.0
certifi==2019.11.28
cffi==1.14.3
chardet==3.0.4
Click==7.0
colorama==0.4.3
cryptography==2.8
cycler==0.10.0
cymem==2.0.3
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
distro==1.4.0
entrypoints==0.3
et-xmlfile==1.0.1
fasteners==0.14.1
filelock==3.0.12
Flask==1.1.2
future==0.18.2
html5lib==1.0.1
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.15.2
jeepney==0.4.3
Jinja2==2.10.1
joblib==0.16.0
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.2
jupyter-console==6.0.0
jupyter-core==4.6.3
jupyterlab==2.2.8
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
keyring==18.0.1
kiwisolver==1.0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
lxml==4.5.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
marshmallow==3.8.0
matplotlib==3.1.2
mistune==0.8.4
monotonic==1.5
more-itertools==4.2.0
murmurhash==1.0.2
nbclient==0.5.0
nbconvert==5.6.1
nbformat==5.0.4
nest-asyncio==1.4.1
netifaces==0.10.4
nltk==3.4.5
notebook==6.0.3
numexpr==2.7.1
numpy==1.17.4
oauthlib==3.1.0
olefile==0.46
openpyxl==3.0.3
packaging==20.3
pandas==0.25.3
pandocfilters==1.4.2
paramiko==2.6.0
parso==0.5.2
patsy==0.5.1
pbr==5.5.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
plac==1.1.3
pluggy==0.13.0
preshed==3.0.2
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.6.1
ptyprocess==0.6.0
py==1.8.1
pycparser==2.20
Pygments==2.3.1
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
pyRFC3339==1.1
pyrsistent==0.15.5
pytest==4.6.9
python-dateutil==2.7.3
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
pyzmq==18.1.1
qtconsole==4.7.7
QtPy==1.9.0
regex==2020.9.27
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
scikit-learn==0.23.2
scipy==1.3.3
seaborn==0.11.0
SecretStorage==2.3.1
Send2Trash==1.5.0
simplejson==3.16.0
six==1.14.0
soupsieve==1.9.5
spacy==2.3.2
srsly==1.0.2
statsmodels==0.12.0
tables==3.6.1
terminado==0.8.2
testpath==0.4.4
testresources==2.0.1
thinc==7.4.1
threadpoolctl==2.1.0
tornado==5.1.1
tqdm==4.50.0
traitlets==4.3.3
urllib3==1.25.8
virtualenv==20.0.31
wadllib==1.3.3
wasabi==0.8.0
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==2.0.0
xlrd==1.1.0
xlwt==1.3.0
zipp==1.0.0
